consumer_three/requirements.txt
Flask==2.2.3
httplib2==0.20.2
mysql-connector==2.2.9
pika==1.3.1
